SevenZip rollback, fixed Lambda and text that I removed by mistake and removed oddloop custom loop string.
