+++ ScoreSaber/Core/Utils/LeaderboardUtils.cs
@@ -6,6 +6,7 @@ using ScoreSaber.Extensions;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 
 #endregion
 
@@ -14,14 +15,8 @@ namespace ScoreSaber.Core.Utils {
 
         internal static bool LocalReplayExists(IDifficultyBeatmap difficultyBeatmap, ScoreMap score) {
 
-            if (File.Exists(GetReplayPath(difficultyBeatmap, score))) {
-                return true;
-            }
-
-            if (File.Exists(GetLegacyReplayPath(difficultyBeatmap, score))) {
-                return true;
-            }
-            return false;
+            return File.Exists(GetReplayPath(difficultyBeatmap, score)) ||
+                   File.Exists(GetLegacyReplayPath(difficultyBeatmap, score));
         }
 
         internal static string GetReplayPath(IDifficultyBeatmap difficultyBeatmap, ScoreMap scoreMap) {
@@ -231,10 +226,8 @@ namespace ScoreSaber.Core.Utils {
 
         public static bool ContainsV3Stuff(IReadonlyBeatmapData readonlyBeatmapData) {
 
-            foreach (var item in readonlyBeatmapData.allBeatmapDataItems)
-                if (item.type == BeatmapDataItem.BeatmapDataItemType.BeatmapObject && item is SliderData)
-                    return true;
-            return false;
+            return readonlyBeatmapData.allBeatmapDataItems.Any(item =>
+                item.type == BeatmapDataItem.BeatmapDataItemType.BeatmapObject && item is SliderData);
         }
     }
 }
