+++ ScoreSaber/Libraries/SevenZip/Compress/LzmaAlone/SevenZipHelper.cs
@@ -2,52 +2,55 @@ using System;
 using System.IO;
 
 
-namespace SevenZip.Compression.LZMAAlone {
-    public static class SevenZipHelper {
+namespace SevenZip.Compression.LZMA
+{
+    public static class SevenZipHelper
+    {
 
-        static int dictionary = 1 << 23;
+       static int dictionary = 1 << 23;
 
-        // static Int32 posStateBits = 2;
-        // static  Int32 litContextBits = 3; // for normal files
+      // static Int32 posStateBits = 2;
+     // static  Int32 litContextBits = 3; // for normal files
         // UInt32 litContextBits = 0; // for 32-bit data
-        // static  Int32 litPosBits = 0;
+     // static  Int32 litPosBits = 0;
         // UInt32 litPosBits = 2; // for 32-bit data
-        // static   Int32 algorithm = 2;
-        // static    Int32 numFastBytes = 128;
+    // static   Int32 algorithm = 2;
+    // static    Int32 numFastBytes = 128;
 
-        static bool eos = false;
-
-
+     static   bool eos = false;
 
 
 
-        static CoderPropID[] propIDs =
-                   {
-                    CoderPropID.DictionarySize,
-                    CoderPropID.PosStateBits,
-                    CoderPropID.LitContextBits,
-                    CoderPropID.LitPosBits,
-                    CoderPropID.Algorithm,
-                    CoderPropID.NumFastBytes,
-                    CoderPropID.MatchFinder,
-                    CoderPropID.EndMarker
-                };
+
+
+     static   CoderPropID[] propIDs = 
+				{
+					CoderPropID.DictionarySize,
+					CoderPropID.PosStateBits,
+					CoderPropID.LitContextBits,
+					CoderPropID.LitPosBits,
+					CoderPropID.Algorithm,
+					CoderPropID.NumFastBytes,
+					CoderPropID.MatchFinder,
+					CoderPropID.EndMarker
+				};
 
         // these are the default properties, keeping it simple for now:
-        static object[] properties =
-                   {
-                    (Int32)(dictionary),
-                    (Int32)(2),
-                    (Int32)(3),
-                    (Int32)(0),
-                    (Int32)(2),
-                    (Int32)(128),
-                    "bt4",
-                    eos
-                };
+     static   object[] properties = 
+				{
+					(Int32)(dictionary),
+					(Int32)(2),
+					(Int32)(3),
+					(Int32)(0),
+					(Int32)(2),
+					(Int32)(128),
+					"bt4",
+					eos
+				};
 
 
-        public static byte[] Compress(byte[] inputBytes) {
+        public static byte[] Compress(byte[] inputBytes)
+        {
 
             MemoryStream inStream = new MemoryStream(inputBytes);
             MemoryStream outStream = new MemoryStream();
@@ -61,11 +64,12 @@ namespace SevenZip.Compression.LZMAAlone {
             return outStream.ToArray();
         }
 
-        public static byte[] Decompress(byte[] inputBytes) {
+        public static byte[] Decompress(byte[] inputBytes)
+        {
             MemoryStream newInStream = new MemoryStream(inputBytes);
 
             SevenZip.Compression.LZMA.Decoder decoder = new SevenZip.Compression.LZMA.Decoder();
-
+            
             newInStream.Seek(0, 0);
             MemoryStream newOutStream = new MemoryStream();
 
@@ -73,7 +77,8 @@ namespace SevenZip.Compression.LZMAAlone {
             if (newInStream.Read(properties2, 0, 5) != 5)
                 throw (new Exception("input .lzma is too short"));
             long outSize = 0;
-            for (int i = 0; i < 8; i++) {
+            for (int i = 0; i < 8; i++)
+            {
                 int v = newInStream.ReadByte();
                 if (v < 0)
                     throw (new Exception("Can't Read 1"));
