--- ScoreSaber/Libraries/SevenZip/Compression/LZ/IMatchFinder.cs
@@ -1,21 +1,0 @@
-// IMatchFinder.cs
-
-using System;
-
-namespace SevenZip.Compression.LZ {
-    interface IInWindowStream {
-        void SetStream(System.IO.Stream inStream);
-        void Init();
-        void ReleaseStream();
-        Byte GetIndexByte(Int32 index);
-        UInt32 GetMatchLen(Int32 index, UInt32 distance, UInt32 limit);
-        UInt32 GetNumAvailableBytes();
-    }
-
-    interface IMatchFinder : IInWindowStream {
-        void Create(UInt32 historySize, UInt32 keepAddBufferBefore,
-                UInt32 matchMaxLen, UInt32 keepAddBufferAfter);
-        UInt32 GetMatches(UInt32[] distances);
-        void Skip(UInt32 num);
-    }
-}
